move raw_stats_converter
